requirements.txt
PyExecJS
redis-py-cluster
requests==2.21.0
pymysql
xlrd
redis
pika
aiohttp==3.7.4.post0
async-timeout==3.0.1
selenium
pdfminer3k
python-docx
demjson
kafka-python==2.0.2
dateparser
jsonpath==0.82
pandas
numpy
confluent_kafka
scrapy
oss2
pdfplumber==0.5.28
ddddocr==1.4.3
egex==2019.8.19